fix: tweak nft empty state
